Added new states and corrected functions that checked whether the user came in contact with any snakes or ladders with if statements
Changed the users movement by only allowing it to move along the x axis.
The user would be sent down to the bottom if in contact with a snake and if the user is in contact with ladders, they can use up and down keys to climb along the ladder to then get closer to the finish. Next I will be adding new mouse is pressed functions to the strings that are newly introduced so that the user can restart if they have won (x amount of times) or if they have lost.