Created separate displays/icons for each function of the recording button
Had to get rid of the original display and switch states and create new function for each icon. Not quite done with this part yet as I will have to finalize mousePressed checks and add the new functions created for each of these icons into the main program to see if it works.